#17 Fix application hanging by making editors load even more lazier
There's a hardcoded limit (10) of threads that can be run simultaneously
using SwingWorker. Previously, having this amount of editors loading at
the same time will cause the application to hang because these threads
are waiting for the navigator nodes to load but there's no an available
slot for a loading job to load these.

By trying really hard, you still can cause the application to hang,
but it's still better than hanging at start.